feat: [1.1.1] Enhance Sadad mobile number regex and improve PHPDoc
Updated the mobile number regex format related to Sadad Payment Service to accept phone numbers starting with 093. Also, addressed and corrected PHPDoc issues identified by PHPStan. Additionally, added phpstan/phpstan for enhanced code analysis and included the new changelog file CHANGELOG.MD.

+++ tests/Unit/Services/PlutuSadad/Traits/ValidFakeDataTrait.php
@@ -46,7 +46,7 @@ trait ValidFakeDataTrait
      */
     protected function createValidFakeBirthYear(): int
     {
-        return (int) rand(1940, date('Y')-18);
+        return (int) rand(1940, (int) date('Y') - 18);
     }
 
     /**
@@ -62,7 +62,7 @@ trait ValidFakeDataTrait
     /**
      * Create a valid fake process Id.
      *
-     * @return float An amount.
+     * @return string An amount.
      */
     protected function createValidFakeProcessId(): string
     {
@@ -73,7 +73,7 @@ trait ValidFakeDataTrait
     /**
      * Create a valid fake OTP.
      *
-     * @return float An amount.
+     * @return string An amount.
      */
     protected function createValidFakeOTP(): string
     {
@@ -83,7 +83,7 @@ trait ValidFakeDataTrait
     /**
      * Create a valid fake invoice number.
      *
-     * @return float An amount.
+     * @return string An amount.
      */
     protected function createValidFakeInvoiceNumber(): string
     {
